+++ src/core/transaction.rs
@@ -108,17 +108,17 @@ pub struct ContractWithdraw<H: Hash, S: SignatureScheme> {
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
 pub struct MpnDeposit<H: Hash, S: SignatureScheme, ZS: ZkSignatureScheme> {
-    pub zk_address_index: u32,
+    pub zk_address_index: u64,
     pub zk_address: ZS::Pub,
-    pub zk_token_index: u32,
+    pub zk_token_index: u64,
     pub payment: ContractDeposit<H, S>,
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
 pub struct MpnWithdraw<H: Hash, S: SignatureScheme, ZS: ZkSignatureScheme> {
-    pub zk_address_index: u32,
+    pub zk_address_index: u64,
     pub zk_address: ZS::Pub,
-    pub zk_token_index: u32,
+    pub zk_token_index: u64,
     pub zk_nonce: u64,
     pub zk_sig: ZS::Sig,
     pub payment: ContractWithdraw<H, S>,
